changed cmd to endpoints
