+++ backend/internal/server/apis/schedule/analysis.go
@@ -1,0 +1,38 @@
+package schedule
+
+import (
+	"fmt"
+	"github.com/gofiber/fiber/v2"
+	"github.com/jak103/powerplay/internal/server/apis"
+	"github.com/jak103/powerplay/internal/server/apis/schedule/pkg/analysis"
+	"github.com/jak103/powerplay/internal/server/apis/schedule/pkg/models"
+	"github.com/jak103/powerplay/internal/server/apis/schedule/pkg/parser"
+	"github.com/jak103/powerplay/internal/server/services/auth"
+	"github.com/jak103/powerplay/internal/utils/responder"
+)
+
+func init() {
+	apis.RegisterHandler(fiber.MethodPost, "/schedule/analysis", auth.Authenticated, handleAnalysis)
+}
+
+func handleAnalysis(c *fiber.Ctx) error {
+	games, seasonConfig := parser.ReadGames("summer_2024")
+
+	analysis.RunTimeAnalysis(games)
+
+	printTeamSchedules(games, seasonConfig)
+	return responder.NotYetImplemented(c)
+}
+
+func printTeamSchedules(games []models.Game, seasonConfig models.SeasonConfig) {
+	for _, league := range seasonConfig.Leagues {
+		for _, team := range league.Teams {
+			fmt.Printf("-----------\n%v\n", team.Name)
+			for _, game := range games {
+				if team.Name == game.Team1Name || team.Name == game.Team2Name {
+					fmt.Printf("%s\n", game)
+				}
+			}
+		}
+	}
+}
