--- backend/internal/server/apis/schedule/cmd/README.md
@@ -1,1 +1,0 @@
-We need to change these into endpoints
