--- backend/internal/server/apis/schedule/schedule.go
@@ -1,3 +1,0 @@
-package schedule
-
-// TODO not sure what this file is for ...
